Add: CreditSystemの一部をProbabilitySystemに分離
