+++ Assets/Scripts/ProbabilitySystem.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: d990c9bf4f2fc02458fb5d2fcf4e421e
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
